// src/events/interfaces/events.enum.ts
export const enum Events {
  PAGE_INFO = 'pageInfo',
  VIEW_STEP = 'viewStep',
  ORIGIN_COMPLETE = 'originComplete',
  DESTINATION_COMPLETE = 'destinationComplete',
  PACKAGES_COMPLETE = 'packagesComplete',
  SHIPMENT_COMPLETE = 'shipmentComplete',
  SHIP_AGAIN = 'shipAgain',
  VIEW_RATE = 'viewRate',
  ERROR = 'error',
}